+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Built inline compact favorites viewer to browse what titles your users have favorited (this is accessible via the ignored favorites block in settings)
+
 ### Fixed
 
 - Candidates and Protected mobile view had some visual bugs